Feat/streaming llm (#219)
* feat: replace FunctionCallValidationError if not exist with AgentThinkAction

* fix: testcases

* fix: lint

* Update: caching mcp tool when starting up server

* Update: Create new streaming_action event

* Update: unit test, caching tools and streaming response

* Update: load mcp tools and description when starting up server

* feat: add fix to tools

* fix: conflict

* chore: merge release

* fix: linting

* feat: add missing dependencies

* fix: params

* fix: compare tool name in mcp

* chore: remove .claude

* chore: precommit lint

* fix: testcases

* Chore: Bypass calculate cost error without custom_provider

* optimize conversation start time & revamp chat mode sysprompt

* Update: Streaming final thought and revert system prompt

* Update: Disable thought of mcp action avoid duplicating message

* chore: use shared config to load tools

* Chore: Update process streaming message llm completion when no tool call

* Update: fix cost streaming

* Update: add tracing span and include usage on streaming mode

* Update: Logging accumulate the token usage and streaming cost

* fix: properly calc prompt caching price for anthropic stream

* fix: add more safe checks for cache read

* fix: pass all tests

* Update: Multiple error to retry mechanism

* chore: Stop streaming final thought when we are in follow up

* Update: Fix unittest with research mode when stream final thought

* chore: add opentelemetry for streaming llm (#224)


* Chore: Update system prompt when user command the unclear message

* Hotfix: Update message action: change wait_for_response

---------

+++ openhands/server/middleware.py
@@ -66,10 +66,10 @@ class CacheControlMiddleware(BaseHTTPMiddleware):
 class InMemoryRateLimiter:
     history: dict
     requests: int
-    seconds: int
-    sleep_seconds: int
+    seconds: float
+    sleep_seconds: float
 
-    def __init__(self, requests: int = 2, seconds: int = 1, sleep_seconds: int = 1):
+    def __init__(self, requests: int = 2, seconds: float = 1, sleep_seconds: float = 1):
         self.requests = requests
         self.seconds = seconds
         self.sleep_seconds = sleep_seconds
